+++ ethereum/contracts/bridge/BridgeGetters.sol
@@ -10,7 +10,11 @@ import "../interfaces/IWormhole.sol";
 import "./BridgeState.sol";
 
 contract BridgeGetters is BridgeState {
-    function governanceActionIsConsumed(bytes32 hash) public view returns (bool) {
+    function governanceActionIsConsumed(bytes32 hash)
+        public
+        view
+        returns (bool)
+    {
         return _state.consumedGovernanceActions[hash];
     }
 
@@ -26,39 +30,43 @@ contract BridgeGetters is BridgeState {
         return IWormhole(_state.wormhole);
     }
 
-    function chainId() public view returns (uint16){
+    function chainId() public view returns (uint16) {
         return _state.provider.chainId;
     }
 
-    function governanceChainId() public view returns (uint16){
+    function governanceChainId() public view returns (uint16) {
         return _state.provider.governanceChainId;
     }
 
-    function governanceContract() public view returns (bytes32){
+    function governanceContract() public view returns (bytes32) {
         return _state.provider.governanceContract;
     }
 
-    function wrappedAsset(uint16 tokenChainId, bytes32 tokenAddress) public view returns (address){
+    function wrappedAsset(uint16 tokenChainId, bytes32 tokenAddress)
+        public
+        view
+        returns (address)
+    {
         return _state.wrappedAssets[tokenChainId][tokenAddress];
     }
 
-    function bridgeContracts(uint16 chainId_) public view returns (bytes32){
+    function bridgeContracts(uint16 chainId_) public view returns (bytes32) {
         return _state.bridgeImplementations[chainId_];
     }
 
-    function tokenImplementation() public view returns (address){
+    function tokenImplementation() public view returns (address) {
         return _state.tokenImplementation;
     }
 
-    function WETH() public view returns (IWETH){
+    function WETH() public view returns (IWETH) {
         return IWETH(_state.provider.WETH);
     }
 
-    function outstandingBridged(address token) public view returns (uint256){
+    function outstandingBridged(address token) public view returns (uint256) {
         return _state.outstandingBridged[token];
     }
 
-    function isWrappedAsset(address token) public view returns (bool){
+    function isWrappedAsset(address token) public view returns (bool) {
         return _state.isWrappedAsset[token];
     }
 }
@@ -66,5 +74,5 @@ contract BridgeGetters is BridgeState {
 interface IWETH is IERC20 {
     function deposit() external payable;
 
-    function withdraw(uint amount) external;
+    function withdraw(uint256 amount) external;
 }
